poolflkadapt is an interface to compute flk / flkadapt on poolseq data.
The input file is a popoolation sync file

For prediction of allele frequencies from counts, the implementation uses
a hierarchical approach, where all pools are used to parameterize a weakly
informative prior then used on each pool independantly.

the prior is beta( a, b)

if pbar < 0.5:
 a =1
 b= (1-pbar)/pbar
else:
 a=pbar/(1-pbar)
 b=1

This can be interpreted as the minimum sample size leading to the observed
mean allele frequency.

Considering this prior, the allele frequency estimate is the mean posterior obtained
by combination with a binomial model for allele counts. So this does not model
haploid pool size but should be more robust to small allele counts.

Current status : calculation of all freqs from a file.

TODO:
implement kinship calculation
implement FLK
implement FLKadapt